Change the Firebase library to support React Native
In a project I'm working on, we depend on `react-native-firebase` and
with this being a different dependency, we need to change the way we
import the database module.

Unfortunately, `react-native-firebase` doesn't have full support for
TypeScript yet and I had to remove some of the typings to get this to
work.